STYLE: move ObjectProxy and ConditionalObject in base.py + update examples
